Playtest III
- Can now scroll in the clipboard screen to cycle between pages
- Fixed schematic checklist putting too many entries on a single page
- Fixed funnels losing filters when changing between types
- Fixed smoothed sloped tracks not working across multiple chained slopes
- Improved location transitions of node locations in-between smoothed sloped tracks
- Fixed track placement allowing an s-bend between two sloped track pieces in specific arrangements
- Updated Crushing Wheel obj file

+++ src/main/java/com/simibubi/create/content/logistics/block/funnel/AbstractFunnelBlock.java
@@ -129,7 +129,8 @@ public abstract class AbstractFunnelBlock extends Block implements IBE<FunnelBlo
 
 	@Override
 	public void onRemove(BlockState state, Level world, BlockPos pos, BlockState newState, boolean isMoving) {
-		IBE.onRemove(state, world, pos, newState);
+		if (state.getBlock() != newState.getBlock() && !isFunnel(newState) || !newState.hasBlockEntity())
+			IBE.onRemove(state, world, pos, newState);
 	}
 
 	@Override
